+++ demos/LawnDart.Demo.Academy.WebApi/Program.cs
@@ -9,6 +9,7 @@ using LawnDart.Authorization.AspNetCore;
 using LawnDart.Demo.Academy.WebApi.Authorization;
 using LawnDart.Demo.Academy.WebApi.Commands;
 using LawnDart.Demo.Academy.WebApi.Domain.Events;
+using LawnDart.Demo.Academy.WebApi.Handlers;
 using LawnDart.Demo.Academy.WebApi.Projections;
 using LawnDart.EventSourcing;
 using LawnDart.EventSourcing.SqlServer;
@@ -28,7 +29,8 @@ builder.Services.AddLawnDart(opts =>
 
 builder.Services.AddInMemoryProjectionStores("default");
 var ctx = builder.Services.AddBoundedContext("default")
-    .WithEventTypes(typeof(StudentRegistered).Assembly);
+    .WithEventTypes<StudentRegistered>()
+    .WithCommandHandlers<RegisterStudentCommandHandler>();
 
 var useSql = args.Contains("--sql")
     || builder.Configuration.GetValue("EventStore:UseSqlServer", false);